+++ src/redisearch_rs/tag_index/tests/integration/main.rs
@@ -16,8 +16,10 @@ extern crate redisearch_rs;
 redis_mock::mock_or_stub_missing_redis_c_symbols!();
 
 mod create;
+mod expansion_timeout;
 mod filtered_iteration;
 mod indexing;
 mod reader;
+mod suffix_wildcard;
 mod tag;
 mod util;
